requirements.txt
futures==3.2.0
mock==2.0.0
PyHamcrest==1.9.0
pytest==3.4.2
